Make TTY::write not add a newline automatically
Then make a bunch of adjustments to the magic numbers to compensate.

What was happening to create the extra newline was the last line was
then scrolling the screen up one, which changed which text was on which
line.

Figured it out when I thought I had lost the ability to count, started
searching around for ansi position problems, and found this:
http://stackoverflow.com/a/9722282/51683

This makes it more idiomatic rust too, since methods like print and
write are assumed to not add a newline, while println and writeln do.

This also changes the meaning of config.visible_limit to mean the number
of results we should display, we take up one more line than that to show
the header line. Some places were assuming that did include the header,
some were assuming it didn't.